core-api: move extension and component data utils into extensions/
