+++ packages/core-api/src/extensions/componentData.test.tsx
