+++ packages/core-api/src/extensions/extensions.test.tsx
